requirements.txt
blinker==1.9.0
build==1.2.2.post1
click==8.1.7
colorama==0.4.6
coverage==7.6.7
Flask==3.1.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
pyproject_hooks==1.2.0
pytest==8.3.3
python-dotenv==1.0.1
Werkzeug==3.1.3
# Editable install with no version control (flaskapp==1.0.0)
-e c:\users\aidan\files\code\financial-formulas